requirements.txt
patool
pandas
numpy
matplotlib